nonfunctional but now with an ubuntu dockerfile with inputs baked in
ENV LANG C.UTF-8 is needed in the Dockerfile with an ubuntu base because for some reason it changes output back into ascii which would throw an error with my unicode output.

this version of python takes ages to build.

this time the error has once again returned `cat: orfalyzer.py: No such file or directory`